% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apify}
\alias{apify}
\title{Enable API for a bamlss model}
\usage{
apify(model, propose = "iwlsC_gp")
}
\arguments{
\item{model}{A \code{\link[bamlss:bamlss]{bamlss::bamlss}} or \code{\link[bamlss:bamlss.frame]{bamlss::bamlss.frame}} object.}

\item{propose}{A string containing the name of a bamlss propose function for
model terms. Defaults to \code{"iwlsC_gp"}. See \code{\link[bamlss:GMCMC]{bamlss::GMCMC()}}.}
}
\value{
The API-enabled model.
}
\description{
This function enables the API for the given bamlss model. Afterwards, you
can use the functions from the bamlssAPI package with the model.
}
\examples{
y <- rnorm(10)
b <- bamlss(y ~ 1)
b <- apify(b)

}
