% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{predictors}
\alias{predictors}
\alias{smooths}
\title{Extract predictor and smooth names}
\usage{
predictors(model)
smooths(model, predictor)
}
\arguments{
\item{model}{An \link[=apify]{API-enabled} bamlss model.}

\item{predictor}{A string containing the name of one predictor in the model.}
}
\value{
A character vector containing either the names of the predictors or
the names of the smooths.
}
\description{
The \code{predictors} function extracts the names of the predictors from the
given bamlss model, while the \code{smooths} function returns the names of the
smooths in one predictor.
}
\examples{
library(bamlss)
y <- rnorm(10)
b <- bamlss(y ~ 1)
b <- apify(b)
predictors(b)
smooths(b, predictor = predictors(b)[1])

}
